Move instance event multiplexers to different file
